/world/rustの削除
