--- frontend/app/worlds/rust/components/profile.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const profile = () => {
-  return (
-    <div>profile</div>
-  )
-}
-
-export default profile
